requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.3.2
pymongo==4.6.0
opencv-python==4.8.1.78
mediapipe==0.10.9
numpy==1.24.3
Pillow==10.1.0
python-jose[cryptography]==3.3.0
structlog==23.2.0
